requirements.txt
flask
tensorflow
Pillow
numpy
opencv-python
gunicorn
flask-cors
